That's it for today.... Need some rest.

+++ x64/src/c/mem.h
@@ -6,46 +6,46 @@
 
 typedef struct __attribute__ ((packed))
 {
-    uint8_t Pr;
-    uint8_t RW;
-    uint8_t USU;
-    uint8_t PWT;
-    uint8_t PCD;
-    uint8_t Acc;
-    uint8_t Sz; //must be 0
+    uint64_t Pr;
+    uint64_t RW;
+    uint64_t USU;
+    uint64_t PWT;
+    uint64_t PCD;
+    uint64_t Acc;
+    uint64_t Sz; //must be 0
     uint64_t base; // Since 4KB-aligned, 12 bits are useless, 52(total) - 12 = 40 bits left
     // will ignore the low 12 bits as well as the high 12 bits of this field
-    uint8_t XD;
+    uint64_t XD;
 } pml4_entry_t, pdpt_entry_t, pd_entry_t;
 
 typedef struct __attribute__ ((packed))
 {
-    uint8_t Pr;
-    uint8_t RW;
-    uint8_t USU;
-    uint8_t PWT;
-    uint8_t PCD;
-    uint8_t Acc;
-    uint8_t dirty;
-    uint8_t PAT;
-    uint8_t Gl;
+    uint64_t Pr;
+    uint64_t RW;
+    uint64_t USU;
+    uint64_t PWT;
+    uint64_t PCD;
+    uint64_t Acc;
+    uint64_t dirty;
+    uint64_t PAT;
+    uint64_t Gl;
     uint64_t base; // Since 4KB-aligned, 12 bits are useless, 52(total) - 12 = 40 bits left
     // will ignore the low 12 bits as well as the high 12 bits of this field
-    uint8_t XD;
+    uint64_t XD;
 } pt_entry_t;
 
 typedef struct __attribute__ ((packed))
 {
-    uint32_t base;
-    uint32_t limit;
-    uint8_t type;
-    uint8_t DPL;
-    uint8_t Gr;
-    uint8_t Acc;
-    uint8_t Pr;
-    uint8_t Sz; //32 bits = 1, 16 bits = 0
-    uint8_t x64;
-    uint8_t Sys; //System = 0, code/data = 1
+    uint64_t base;
+    uint64_t limit;
+    uint64_t type;
+    uint64_t DPL;
+    uint64_t Gr;
+    uint64_t Acc;
+    uint64_t Pr;
+    uint64_t Sz; //32 bits = 1, 16 bits = 0
+    uint64_t x64;
+    uint64_t Sys; //System = 0, code/data = 1
 } segment_descriptor_t;
 
 typedef struct __attribute__ ((packed))
@@ -57,7 +57,7 @@ typedef struct __attribute__ ((packed))
 typedef struct __attribute__ ((packed))
 {
     uint16_t limit;
-    uint32_t base;
+    uint64_t base;
 } idt_ptr_t;
 
 void HYPKERNEL64 hk_write_segment_descriptor(uint8_t *const gdt, segment_descriptor_t const *const seg_desc);
