+++ x86/src/c/mem.h
@@ -5,71 +5,71 @@
 
 typedef struct __attribute__ ((packed))
 {
-    uint8_t Pr;
-    uint8_t RW;
-    uint8_t USU;
-    uint8_t PWT;
-    uint8_t PCD;
-    uint8_t Acc;
-    uint8_t Sz; //must be 0
+    uint32_t Pr;
+    uint32_t RW;
+    uint32_t USU;
+    uint32_t PWT;
+    uint32_t PCD;
+    uint32_t Acc;
+    uint32_t Sz; //must be 0
     uint64_t base; // Since 4KB-aligned, 12 bits are useless, 52(total) - 12 = 40 bits left
                  // will ignore the low 12 bits as well as the high 12 bits of this field
-                 uint8_t XD;
+                 uint32_t XD;
 } pml4_entry_t, pdpt_entry_t, pd_entry_t;
 
 typedef struct __attribute__ ((packed))
 {
-    uint8_t Pr;
-    uint8_t RW;
-    uint8_t USU;
-    uint8_t PWT;
-    uint8_t PCD;
-    uint8_t Acc;
-    uint8_t dirty;
-    uint8_t PAT;
-    uint8_t Gl;
+    uint32_t Pr;
+    uint32_t RW;
+    uint32_t USU;
+    uint32_t PWT;
+    uint32_t PCD;
+    uint32_t Acc;
+    uint32_t dirty;
+    uint32_t PAT;
+    uint32_t Gl;
     uint64_t base; // Since 4KB-aligned, 12 bits are useless, 52(total) - 12 = 40 bits left
     // will ignore the low 12 bits as well as the high 12 bits of this field
-    uint8_t XD;
+    uint32_t XD;
 } pt_entry_t;
 
 typedef struct __attribute__ ((packed))
 {
     uint32_t offset;
-    uint16_t seg_sel;
-    uint8_t Pr;
-    uint8_t DPL;
-    uint8_t Sz;
+    uint32_t seg_sel;
+    uint32_t Pr;
+    uint32_t DPL;
+    uint32_t Sz;
 } interrupt_gate_t;
 
 typedef struct __attribute__ ((packed))
 {
     uint32_t offset;
-    uint16_t seg_sel;
-    uint8_t Pr;
-    uint8_t DPL;
-    uint8_t Sz;
+    uint32_t seg_sel;
+    uint32_t Pr;
+    uint32_t DPL;
+    uint32_t Sz;
 } trap_gate_t;
 
 typedef struct __attribute__ ((packed))
 {
-    uint16_t tss_sel;
-    uint8_t DPL;
-    uint8_t Pr;
+    uint32_t tss_sel;
+    uint32_t DPL;
+    uint32_t Pr;
 } task_gate_t;
 
 typedef struct __attribute__ ((packed))
 {
     uint32_t base;
     uint32_t limit;
-    uint8_t type;
-    uint8_t DPL;
-    uint8_t Gr;
-    uint8_t Acc;
-    uint8_t Pr;
-    uint8_t Sz; //32 bits = 1, 16 bits = 0
-    uint8_t x64;
-    uint8_t Sys; //System = 0, code/data = 1
+    uint32_t type;
+    uint32_t DPL;
+    uint32_t Gr;
+    uint32_t Acc;
+    uint32_t Pr;
+    uint32_t Sz; //32 bits = 1, 16 bits = 0
+    uint32_t x64;
+    uint32_t Sys; //System = 0, code/data = 1
 
 } segment_descriptor_t;
